Rename py.global to py.Dynamic.global
