+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Dynamic.scala
@@ -6,6 +6,26 @@ import me.shadaj.scalapy.interpreter.CPythonInterpreter
 
 @native trait Dynamic extends Any with AnyDynamics
 
+object Dynamic {
+  object global extends scala.Dynamic {
+    def applyDynamic(method: String)(params: Any*): Dynamic = {
+      Any.populateWith(CPythonInterpreter.callGlobal(method, params.map(_.value), Seq())).as[Dynamic]
+    }
+
+    def applyDynamicNamed(method: String)(params: (String, Any)*): Dynamic = {
+      Any.populateWith(CPythonInterpreter.callGlobal(
+        method,
+        params.filter(_._1.isEmpty).map(_._2.value),
+        params.filter(_._1.nonEmpty).map(t => (t._1, t._2.value))
+      )).as[Dynamic]
+    }
+
+    def selectDynamic(value: String): Dynamic = {
+      Any.populateWith(CPythonInterpreter.selectGlobal(value)).as[Dynamic]
+    }
+  }
+}
+
 trait AnyDynamics extends scala.Any with Any with scala.Dynamic {
   def applyDynamic(method: String)(params: Any*): Dynamic = {
     Any.populateWith(CPythonInterpreter.call(value, method, params.map(_.value), Seq())).as[Dynamic]
